cool

+++ python/ffibuilder.py
@@ -386,7 +386,7 @@ struct rsyscall_bootstrap {
     int listening_sock;
     int syscall_sock;
     int data_sock;
-    int envp_count;
+    size_t envp_count;
 };
 struct rsyscall_stdin_bootstrap {
     struct rsyscall_symbol_table symbols;
@@ -394,7 +394,7 @@ struct rsyscall_stdin_bootstrap {
     int syscall_fd;
     int data_fd;
     int futex_memfd;
-    int envp_count;
+    size_t envp_count;
 };
 
 #define SYS_sendmsg ...
